resolveMutation => resolve-mutation.ts
